facetimehd: Disable ASPM
Kernel commit: 5a8361f7ecceaed64b4064000d16cb703462be49
ACPICA: Integrate package handling with module-level code

The above commit fixed a problem with ACPI that previously prevented
ASPM to be enabled. Unfortunately our driver broke because the link to
the device went down while streaming. Fix this for now by disabling ASPM
for the device entirely.

Note: There is an ACPI method we're calling to "power on/off" the device
so the real solution here might be to get it working correctly. I will
need to investigate that further so no ASPM will have to do for now.

Signed-off-by: Patrik Jakobsson <patrik.r.jakobsson@gmail.com>