return 0 instead of -EINVAL when no filename found
My dev_priv->sensorid1 value is 0x248, which doesn't appear in the case statement. That said, the code below my change indicates that a set file itself is optional, so there should be no difference in behavior if the filename is not found or the filename is known but doesn't exist. The module and camera are functional on my machine with this change, with the caveat that no calibration is performed.