Attach fthd_pci_remove() to pci driver shutdown hook
When running Linux in a VM with the facetimehd module, the device must explicitly be removed.

If the device is not removed upon a shutdown of the VM, the device becomes inoperable until a full reboot of the host. Inoperable here means that the error message "Init failed! No wake signal" occurs.